requirements.txt
pillow==9.5.0
customtkinter==5.2.1
